refactor: split ConversationEngine into focused modules (closes #65)
engine.py (530 lines) was doing too much. Split into:
- engine.py (191 lines) — orchestrator only
- analysis_loop.py (150 lines) — AnalysisLoop, AnalysisResult, format_evidence
- synthesizer.py (169 lines) — ResponseSynthesizer, ComparisonSynthesizer
- dispatcher.py (59 lines) — invoke_tool, invoke_tools

Each module now has a single responsibility. All existing tests pass
with updated import paths.
