+++ src/tracer/conversation/__init__.py
@@ -1,17 +1,8 @@
-from tracer.conversation.engine import (
-    AnalysisLoop,
-    AnalysisResult,
-    ComparisonSynthesizer,
-    ConversationEngine,
-    EngineResponse,
-    ResponseSynthesizer,
-)
-from tracer.conversation.intent import (
-    INTENT_TOOL_MAP,
-    Intent,
-    IntentParser,
-    IntentType,
-)
+from tracer.conversation.analysis_loop import AnalysisLoop, AnalysisResult
+from tracer.conversation.dispatcher import invoke_tool, invoke_tools
+from tracer.conversation.engine import ConversationEngine, EngineResponse
+from tracer.conversation.intent import INTENT_TOOL_MAP, Intent, IntentParser, IntentType
+from tracer.conversation.synthesizer import ComparisonSynthesizer, ResponseSynthesizer
 
 __all__ = [
     "AnalysisLoop",
@@ -24,4 +15,6 @@ __all__ = [
     "IntentParser",
     "IntentType",
     "ResponseSynthesizer",
+    "invoke_tool",
+    "invoke_tools",
 ]
