+++ src/tracer/conversation/dispatcher.py
@@ -1,0 +1,59 @@
+"""Tool dispatcher — routes intent to pipeline tool functions."""
+
+from __future__ import annotations
+
+import asyncio
+
+from tracer.conversation.intent import Intent
+from tracer.data.registry import DataRegistry
+from tracer.models import ToolResult
+from tracer.tools import pipeline
+
+# Maps tool name → the kind of argument it expects.
+TOOL_DISPATCH: dict[str, str] = {
+    "price_event": "ticker",
+    "news": "ticker",
+    "insider": "ticker",
+    "macro": "indicator",
+    "fundamentals": "ticker",
+    "cross_market": "tickers",
+    "memory_search": "query",
+}
+
+
+async def invoke_tool(tool_name: str, intent: Intent, registry: DataRegistry) -> ToolResult:
+    """Invoke a single pipeline tool based on the intent context."""
+    tickers = intent.tickers
+
+    if tool_name == "price_event" and tickers:
+        return await pipeline.price_event(tickers[0], registry)
+    if tool_name == "news" and tickers:
+        return await pipeline.news(tickers[0], registry)
+    if tool_name == "insider" and tickers:
+        return await pipeline.insider(tickers[0], registry)
+    if tool_name == "fundamentals" and tickers:
+        return await pipeline.fundamentals(tickers[0], registry)
+    if tool_name == "cross_market" and tickers:
+        return await pipeline.cross_market(tickers, registry)
+    if tool_name == "macro":
+        return await pipeline.macro(intent.raw_query, registry)
+    if tool_name == "memory_search":
+        return await pipeline.memory_search(intent.raw_query)
+
+    return ToolResult(
+        tool=tool_name,
+        success=False,
+        data={},
+        source="dispatcher",
+        error=f"Cannot invoke {tool_name}: no tickers in query",
+    )
+
+
+async def invoke_tools(
+    tool_names: list[str], intent: Intent, registry: DataRegistry
+) -> list[ToolResult]:
+    """Invoke multiple pipeline tools concurrently."""
+    if not tool_names:
+        return []
+    coros = [invoke_tool(name, intent, registry) for name in tool_names]
+    return list(await asyncio.gather(*coros))
